backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dmitriy Mikhalev <fakeforall@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
psycopg2-binary = "^2.9.9"
lxml = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
